# packages/api-zod/src/generated/types/block.ts
/**
 * Generated by orval v8.13.0 🍺
 * Do not edit manually.
 * Api
 * LegacyBTC Mining Pool API
 * OpenAPI spec version: 0.1.0
 */
import type { BlockMode } from './blockMode';

export interface Block {
  id: number;
  height: number;
  hash: string;
  reward: number;
  finderAddress: string;
  foundAt: Date;
  confirmed: boolean;
  mode: BlockMode;
  devFee: number;
  /** @nullable */
  txid?: string | null;
}