# packages/api-zod/src/generated/types/payout.ts
/**
 * Generated by orval v8.13.0 🍺
 * Do not edit manually.
 * Api
 * LegacyBTC Mining Pool API
 * OpenAPI spec version: 0.1.0
 */

export interface Payout {
  id: number;
  address: string;
  amount: number;
  /** @nullable */
  txid?: string | null;
  createdAt: Date;
  /** @nullable */
  blockId?: number | null;
}