# packages/api-zod/src/generated/types/poolStats.ts
/**
 * Generated by orval v8.13.0 🍺
 * Do not edit manually.
 * Api
 * LegacyBTC Mining Pool API
 * OpenAPI spec version: 0.1.0
 */

export interface PoolStats {
  /** Pool hashrate in H/s */
  hashrate: number;
  activeMiners: number;
  activeWorkers: number;
  totalBlocksFound: number;
  soloBlocksFound: number;
  pplnsBlocksFound: number;
  networkDifficulty: number;
  networkHashrate: number;
  blockHeight: number;
  /** @nullable */
  lastBlockFoundAt: Date | null;
  /** Total LBTC paid out to miners */
  totalPaidOut: number;
}